pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finviz-mcp-server"
version = "1.0.0"
description = "Advanced stock screening and financial analysis MCP server with Finviz integration for AI-powered investment research"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "TraderMonty", email = "noreply@finviz-mcp.com"}
]
keywords = [
    "mcp", "finviz", "stocks", "finance", "screening", "investment", 
    "trading", "analysis", "earnings", "technical-analysis", "fundamental-analysis",
    "claude", "ai", "model-context-protocol", "sec-filings", "market-data"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Operating System :: OS Independent",
    "Environment :: Console",
]

dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/tradermonty/finviz-mcp-server"
Repository = "https://github.com/tradermonty/finviz-mcp-server"
Issues = "https://github.com/tradermonty/finviz-mcp-server/issues"
Documentation = "https://github.com/tradermonty/finviz-mcp-server/blob/main/docs/"

[project.scripts]
finviz-mcp-server = "src.server:cli_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
src = ["py.typed", "*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "live: marks tests as live integration tests (require API key)",
    "slow: marks tests as slow-running",
    "unit: marks tests as unit tests (no external dependencies)",
]